+++ sdk/core/core-client-rest/test/node/streams.spec.ts
@@ -74,12 +74,12 @@ describe("[Node] Streams", () => {
     }
   });
 
-  it("should be able to return a stream for SSEs", async () => {
+  it("should be able to return a stream if there is one", async () => {
     const client = getClient(mockBaseUrl, {
       httpClient: {
         sendRequest: async (request) => {
           return {
-            headers: createHttpHeaders({ "content-type": "text/event-stream" }),
+            headers: createHttpHeaders(),
             status: 200,
             request,
             readableStreamBody: new PassThrough(),