+++ sdk/openai/openai/sources/customizations/api/oaiSse.ts
@@ -2,16 +2,20 @@
 // Licensed under the MIT license.
 
 import { StreamableMethod } from "@azure-rest/core-client";
-import { getSSEs } from "./getSSEs.js";
 import { wrapError } from "./util.js";
+import { toSSE } from "./sse.js";
+import { isUnexpected } from "../../generated/src/rest/isUnexpected.js";
 
 export async function* getOaiSSEs<TEvent>(
-  response: StreamableMethod<unknown>,
+  responseObj: StreamableMethod,
   toEvent: (obj: Record<string, any>) => TEvent
 ): AsyncIterable<TEvent> {
-  const stream = await getSSEs(response);
   let isDone = false;
-  for await (const event of stream) {
+  const response = await responseObj;
+  if (isUnexpected(response)) {
+    throw response.body.error;
+  }
+  for await (const event of toSSE(response.body)) {
     if (isDone) {
       // handle a case where the service sends excess stream
       // data after the [DONE] event