+++ sdk/core/core-rest-pipeline/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Other Changes
 
 - Set `init.duplex` to `"half"` when streaming body via `fetch()` [PR #26890](https://github.com/Azure/azure-sdk-for-js/pull/26890)
+- Return a stream if the content type is text/event-stream.
 
 ## 1.12.0 (2023-08-08)