+++ sdk/openai/openai/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Handles errors from the server more gracefully in the `list`-prefixed operations.
+
 ## 1.0.0-beta.5 (2023-08-25)
 
 ### Features Added